requirements.txt
openai>=1.30.0
anthropic>=0.40.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0
pydub>=0.25.1
google-genai>=1.0.0
